// src/API.ts
export type ListPostsQuery = {
  listPosts:  {
    __typename: "ModelPostConnection",
    items:  Array< {
      __typename: "Post",
      id: string,
      postOwnerId: string,
      postOwnerUsername: string,
      postTitle: string,
      postBody: string,
      createdAt: string | null,
      comments:  {
        __typename: "ModelCommentConnection",
        nextToken: string | null,
      } | null,
      likes:  {
        __typename: "ModelLikeConnection",
        nextToken: string | null,
      } | null,
      updatedAt: string,
    } | null > | null,
    nextToken: string | null,
  } | null,
};